requirements.txt
appnope==0.1.0
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
Django==1.8
djangorestframework==3.3.2
elasticsearch==5.5.1
elasticsearch-dsl==5.4.0
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
mysqlclient==1.3.12
oauthlib==2.0.6
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
tweepy==3.5.0
urllib3==1.21.1
wcwidth==0.1.7
